fix capturing video on Windows

+++ trunk/win/capture.h
@@ -60,8 +60,8 @@ public:
     virtual bool set_buffer(unsigned char* buf, out_format fmt);
     HRESULT async_read()const;
 
-    void decode_to_buffer(unsigned char* src, unsigned int length)const;
-    void decode_padded_to_buffer(unsigned char* scanline0, unsigned int pitch)const;
+    void decode_to_buffer(unsigned char* src, unsigned int length) const;
+    void decode_padded_to_buffer(unsigned char* scanline0, long pitch) const;
 
     static unsigned int enum_devices(char buffers[][128], const unsigned int size)throw();
